sandbox - use electron directly and not preload for postMessage calls
