+++ src/vs/platform/ipc/electron-sandbox/mainProcessService.ts
@@ -3,11 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IChannel, IServerChannel, StaticRouter } from 'vs/base/parts/ipc/common/ipc';
+import { IChannel, IServerChannel } from 'vs/base/parts/ipc/common/ipc';
 import { Client as IPCElectronClient } from 'vs/base/parts/ipc/electron-sandbox/ipc.electron';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
-import { Server as MessagePortServer } from 'vs/base/parts/ipc/electron-sandbox/ipc.mp';
 
 export const IMainProcessService = createDecorator<IMainProcessService>('mainProcessService');
 
@@ -43,26 +42,5 @@ export class ElectronIPCMainProcessService extends Disposable implements IMainPr
 
 	registerChannel(channelName: string, channel: IServerChannel<string>): void {
 		this.mainProcessConnection.registerChannel(channelName, channel);
-	}
-}
-
-/**
- * An implementation of `IMainProcessService` that leverages MessagePorts.
- */
-export class MessagePortMainProcessService implements IMainProcessService {
-
-	declare readonly _serviceBrand: undefined;
-
-	constructor(
-		private server: MessagePortServer,
-		private router: StaticRouter
-	) { }
-
-	getChannel(channelName: string): IChannel {
-		return this.server.getChannel(channelName, this.router);
-	}
-
-	registerChannel(channelName: string, channel: IServerChannel<string>): void {
-		this.server.registerChannel(channelName, channel);
 	}
 }
